src/requirements.txt
certifi==2018.10.15
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.7
mathparse==0.1.2
nltk==3.4.5
oauthlib==2.1.0
Pillow==6.2.0
PyAutoGUI==0.9.47
pymongo==3.7.2
python-dateutil==2.6.1
python-twitter==3.5
pyyaml>=4.2b1
requests==2.20.1
requests-oauthlib==1.0.0
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.3.0
urllib3==1.24.2
wikipedia==1.4.0
wget==3.2